- Added option to use DeferredRenderer without tessellation
- Added non-tessellation shaders for Terrain
- Added Lod plane to ShapeCreator
- Fixed small bugs with PerlinNoise
- Fixed some issues with TerrainChunk (WIP)

+++ code/include/Core/Math/PerlinNoise.h
@@ -123,8 +123,8 @@ namespace Pu
 
 		static float Fade(float t);
 
-		float Gradient(size_t i, float x) const;
-		float Gradient(size_t i, float x, float y) const;
-		float Gradient(size_t i, float x, float y, float z) const;
+		float Gradient(uint32 i, float x) const;
+		float Gradient(uint32 i, float x, float y) const;
+		float Gradient(uint32 i, float x, float y, float z) const;
 	};
 }
